Optimized Eye component
commented out the rinnegan because I didn't like the look of it so I doubled down on the sharingan eyes which looks better. I will build the mask that it goes into.

Small changes also include splitting the feature component, one for home page which is pretty. much good to go, one for portfolio. Not fully designed (old)